+++ packages/eyes-sdk-core/lib/errors/CoordinatesTypeConversionError.js
@@ -1,6 +1,6 @@
 'use strict'
 
-const {EyesError} = require('@applitools/eyes-common')
+const {EyesError} = require('./EyesError')
 
 /**
  * Encapsulates an error converting between two coordinate types.