+++ packages/eyes-sdk-core/lib/frames/FrameChain.js
@@ -1,12 +1,12 @@
 'use strict'
 
-const {ArgumentGuard, Location} = require('@applitools/eyes-common')
+const {ArgumentGuard, Location} = require('../..')
 const {Frame} = require('./Frame')
 const {NoFramesError} = require('../errors/NoFramesError')
 /**
- * @typedef {import('@applitools/eyes-common').Logger} Logger
- * @typedef {import('@applitools/eyes-common').Location} Location
- * @typedef {import('@applitools/eyes-common').RectangleSize} RectangleSize
+ * @typedef {import('../..').Logger} Logger
+ * @typedef {import('../..').Location} Location
+ * @typedef {import('../..').RectangleSize} RectangleSize
  */
 
 class FrameChain {