+++ packages/eyes-sdk-core/lib/positioning/ScrollPositionMemento.js
@@ -1,6 +1,6 @@
 'use strict'
 
-const {Location} = require('@applitools/eyes-common')
+const {Location} = require('../..')
 const PositionMemento = require('./PositionMemento')
 
 /**