+++ packages/sdk-shared/src/sdk-shared.js
@@ -1,3 +1,0 @@
-console.log(
-  `This file is here just so "src" isn't an empty folder. Remove it once you add anything`,
-)