+++ packages/visual-grid-client/CHANGELOG.md
@@ -4,6 +4,10 @@
 
 
 
+## 14.2.0 - 2020/5/24
+
+- updated to @applitools/eyes-sdk-core@10.3.0
+
 ## 14.1.0 - 2020/5/19
 
 - support accessibilitySettings